+++ R/plot.netrank.R
@@ -616,7 +616,7 @@ plot.netrank <- function(...,
   
   plt <- plt + theme_classic()
   
-  plt <- plt + geom_tile() + xlab("") + ylab("") +
+  plt <- plt + geom_tile() + ggplot2::xlab("") + ggplot2::ylab("") +
     theme(line = element_blank(),
           axis.ticks = element_blank(),
           axis.text.x = element_text(size = axis.size, face = axis.face,